requirements.txt
# Minimal requirements for Cloud Run deployment
# Only core API Gateway functionality

# API server
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Configuration management  
python-dotenv

# HTTP requests for inter-service communication
requests>=2.31.0
httpx>=0.27.0
PyPDF2>=3.0.0

# Enhanced logging
colorlog>=6.7.0

# Cloud storage for persistent data
google-cloud-storage>=2.10.0

# Basic dependencies only - no heavy ML libraries
# ChromaDB, FAISS, sentence-transformers moved to microservices
# CRAG evaluation moved to Document Processing Service